Move app building to a separate function and tidy up index file
